enterprises/requirements.txt
asgiref==3.4.1
Django==3.0.7
django-filter==2.3.0
pytz==2021.3
sqlparse==0.4.2
typing-extensions==3.10.0.2
